feat: add relations between User and Todo models

+++ routes/modules/records.js
@@ -13,17 +13,18 @@ router.get('/new', (req, res) => {
 
 // post created data
 router.post('/', (req, res) => {
-  // const user = req.user
+  const userId = req.user._id
   const { name, date, category, amount } = req.body
-  return Record.create({ name, date, category, amount })
+  return Record.create({ name, date, category, amount, userId })
     .then(() => res.redirect('/'))
     .catch((error) => console.log(error))
 })
 
 // show edit page
 router.get('/:id/edit', (req, res) => {
+  const userId = req.user._id
   const _id = req.params.id
-  return Record.findOne({ _id })
+  return Record.findOne({ _id, userId })
     .lean()
     .then((record) => res.render('edit', { record }))
     .catch(error => console.log(error))
@@ -31,11 +32,12 @@ router.get('/:id/edit', (req, res) => {
 
 // update edite data
 router.put('/:id', (req, res) => {
+  const userId = req.user._id
   const _id = req.params.id
   const { name, date, category, amount } = req.body
   // findOneAndUpdate:接受两个参数：一个查询条件的对象和一个包含要更新的字段和值的对象
   // { new: true }: 为了确保返回更新后的文档而不是更新前的文档
-  return Record.findOneAndUpdate({ _id }, { name, date, category, amount }, { new: true })
+  return Record.findOneAndUpdate({ _id, userId }, { name, date, category, amount }, { new: true })
     .lean()
     .then(() => res.redirect('/'))
     .catch(error => console.log(error))
@@ -43,8 +45,9 @@ router.put('/:id', (req, res) => {
 
 // delete record
 router.delete('/:id', (req, res) => {
-  const _id = req.params.id // 為什麼不能用 req.params.＿id
-  return Record.findOneAndDelete({ _id })
+  const userId = req.user._id
+  const _id = req.params.id
+  return Record.findOneAndDelete({ _id, userId })
     .lean()
     // .then((record) => record.remove())
     .then(() => res.redirect('/'))
